rename files from auth folder, rule vue name = name files
