+++ frontend/src/components/Auth/AuthCarousel.vue
