+++ frontend/src/components/Auth/AuthFooter.vue
