+++ frontend/src/components/Auth/AuthNavbar.vue
