+++ frontend/src/layouts/AuthLayout.spec.js
@@ -43,6 +43,10 @@ describe('AuthLayout', () => {
       it('has Component AuthMobileStart', () => {
         expect(wrapper.findComponent({ name: 'AuthMobileStart' }).exists()).toBe(true)
       })
+
+      it('has Component AuthNavbarSmall', () => {
+        expect(wrapper.findComponent({ name: 'AuthNavbarSmall' }).exists()).toBe(true)
+      })
     })
 
     describe('Desktop Version Start', () => {
